# openapi.json
{"openapi":"3.0.0","info":{"title":"Nexus API","version":"0.1.0"},"paths":{"/authors":{"get":{"tags":["Authors"],"summary":"List authors","parameters":[{"name":"page","in":"query","description":"Page index (1-based)","schema":{"type":"integer","format":"int64"}},{"name":"pageSize","in":"query","description":"Page size","schema":{"type":"integer","format":"int64"}},{"name":"sort","in":"query","description":"Sort fields (e.g. lastSeen:desc)","schema":{"type":"array","items":{"type":"string"}}},{"name":"q","in":"query","description":"Search term for name or email","schema":{"type":"string"}},{"name":"listSlug","in":"query","description":"Filter by mailing list slug","schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthorList"}}}}}}},"/authors/{author_id}":{"get":{"tags":["Authors"],"summary":"Get author","parameters":[{"name":"author_id","in":"path","description":"Author id","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthorResponse"}}}}}}},"/authors/{author_id}/lists/{slug}/emails":{"get":{"tags":["Authors"],"summary":"List author emails","parameters":[{"name":"author_id","in":"path","description":"Author id","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"slug","in":"path","description":"Mailing list slug","required":true,"schema":{"type":"string"}},{"name":"page","in":"query","description":"Page index (1-based)","schema":{"type":"integer","format":"int64"}},{"name":"pageSize","in":"query","description":"Page size","schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthorEmailsResponse"}}}}}}},"/authors/{author_id}/lists/{slug}/threads-participated":{"get":{"tags":["Authors"],"summary":"List threads participated by author","parameters":[{"name":"author_id","in":"path","description":"Author id","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"slug","in":"path","description":"Mailing list slug","required":true,"schema":{"type":"string"}},{"name":"page","in":"query","description":"Page index (1-based)","schema":{"type":"integer","format":"int64"}},{"name":"pageSize","in":"query","description":"Page size","schema":{"type":"integer","format":"int64"}},{"name":"sort","in":"query","description":"Sort fields (e.g. lastActivity:desc)","schema":{"type":"array","items":{"type":"string"}}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthorThreadsResponse"}}}}}}},"/authors/{author_id}/lists/{slug}/threads-started":{"get":{"tags":["Authors"],"summary":"List threads started by author","parameters":[{"name":"author_id","in":"path","description":"Author id","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"slug","in":"path","description":"Mailing list slug","required":true,"schema":{"type":"string"}},{"name":"page","in":"query","description":"Page index (1-based)","schema":{"type":"integer","format":"int64"}},{"name":"pageSize","in":"query","description":"Page size","schema":{"type":"integer","format":"int64"}},{"name":"sort","in":"query","description":"Sort fields (e.g. lastActivity:desc)","schema":{"type":"array","items":{"type":"string"}}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthorThreadsResponse"}}}}}}},"/docs":{"get":{"tags":["Docs"],"summary":"API docs (RapiDoc)","responses":{"200":{"description":"OK"}}}},"/health":{"get":{"tags":["Health"],"summary":"Health check","responses":{"200":{"description":"OK"}}}},"/hello":{"get":{"tags":["Health"],"summary":"Hello endpoint","responses":{"200":{"description":"OK"}}}},"/jobs":{"get":{"tags":["Jobs"],"summary":"List jobs","parameters":[{"name":"status","in":"query","description":"Filter jobs by status (queued/running/succeeded/failed)","schema":{"type":"string"}},{"name":"limit","in":"query","description":"Max results","schema":{"type":"integer","format":"int64"}},{"name":"offset","in":"query","description":"Offset into result set","schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobList"}}}}}},"post":{"tags":["Jobs"],"summary":"Enqueue job","requestBody":{"description":"Job payload and scheduling options","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EnqueueRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EnqueueResponse"}}}}}}},"/jobs/stats":{"get":{"tags":["Jobs"],"summary":"Job statistics","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobStats"}}}}}}},"/jobs/{id}":{"get":{"tags":["Jobs"],"summary":"Get job status","parameters":[{"name":"id","in":"path","description":"Job id","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Job"}}}}}}},"/lists/stats":{"get":{"tags":["Lists"],"summary":"Aggregate list stats","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AggregateStatsResponse"}}}}}}},"/lists/{slug}/emails":{"get":{"tags":["Emails"],"summary":"List emails","parameters":[{"name":"slug","in":"path","description":"Mailing list slug","required":true,"schema":{"type":"string"}},{"name":"page","in":"query","description":"Page index (1-based)","schema":{"type":"integer","format":"int64"}},{"name":"pageSize","in":"query","description":"Page size","schema":{"type":"integer","format":"int64"}},{"name":"sort","in":"query","description":"Sort fields (e.g. date:desc)","schema":{"type":"array","items":{"type":"string"}}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmailList"}}}}}}},"/lists/{slug}/emails/{email_id}":{"get":{"tags":["Emails"],"summary":"Get email","parameters":[{"name":"slug","in":"path","description":"Mailing list slug","required":true,"schema":{"type":"string"}},{"name":"email_id","in":"path","description":"Email id","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmailResponse"}}}}}}},"/lists/{slug}/stats":{"get":{"tags":["Lists"],"summary":"Mailing list stats","parameters":[{"name":"slug","in":"path","description":"Mailing list slug","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MailingListStatsResponse"}}}}}}},"/lists/{slug}/threads":{"get":{"tags":["Threads"],"summary":"List threads","parameters":[{"name":"slug","in":"path","description":"Mailing list slug","required":true,"schema":{"type":"string"}},{"name":"page","in":"query","description":"Page index (1-based)","schema":{"type":"integer","format":"int64"}},{"name":"pageSize","in":"query","description":"Page size","schema":{"type":"integer","format":"int64"}},{"name":"sort","in":"query","description":"Sort fields (e.g. lastActivity:desc)","schema":{"type":"array","items":{"type":"string"}}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ThreadList"}}}}}}},"/lists/{slug}/threads/{thread_id}":{"get":{"tags":["Threads"],"summary":"Get thread","parameters":[{"name":"slug","in":"path","description":"Mailing list slug","required":true,"schema":{"type":"string"}},{"name":"thread_id","in":"path","description":"Thread id","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ThreadDetailResponse"}}}}}}},"/mailing-lists":{"get":{"tags":["MailingLists"],"summary":"List mailing lists","parameters":[{"name":"enabled","in":"query","description":"Filter enabled lists","schema":{"type":"boolean"}},{"name":"limit","in":"query","description":"Max results","schema":{"type":"integer","format":"int64"}},{"name":"offset","in":"query","description":"Offset into result set","schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MailingListList"}}}}}}},"/mailing-lists/enabled":{"get":{"tags":["MailingLists"],"summary":"List enabled mailing lists","parameters":[{"name":"limit","in":"query","description":"Max results","schema":{"type":"integer","format":"int64"}},{"name":"offset","in":"query","description":"Offset into result set","schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MailingListList"}}}}}}},"/mailing-lists/{slug}":{"patch":{"tags":["MailingLists"],"summary":"Update mailing list","parameters":[{"name":"slug","in":"path","description":"Mailing list slug","required":true,"schema":{"type":"string"}}],"requestBody":{"description":"Mailing list update payload","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateMailingListRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MailingList"}}}}}}},"/openapi.json":{"get":{"tags":["Docs"],"summary":"OpenAPI spec","responses":{"200":{"description":"OK"}}}},"/webhooks/grokmirror":{"post":{"tags":["Webhooks"],"summary":"Grokmirror webhook","parameters":[{"name":"x-nexus-webhook-secret","in":"header","description":"Shared secret for webhook authentication","schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WebhookEnqueueResponse"}}}}}}}},"components":{"schemas":{"AggregateStatsResponse":{"title":"ApiResponse_for_ListAggregateStats","description":"Standard response envelope used by explorer endpoints.","type":"object","required":["data"],"properties":{"data":{"$ref":"#/definitions/ListAggregateStats"},"meta":{"default":{},"allOf":[{"$ref":"#/definitions/ResponseMeta"}]}}},"AuthorEmailsResponse":{"title":"ApiResponse_for_Array_of_EmailWithAuthor","description":"Standard response envelope used by explorer endpoints.","type":"object","required":["data"],"properties":{"data":{"type":"array","items":{"$ref":"#/definitions/EmailWithAuthor"}},"meta":{"default":{},"allOf":[{"$ref":"#/definitions/ResponseMeta"}]}}},"AuthorList":{"title":"ApiResponse_for_Array_of_AuthorWithStats","description":"Standard response envelope used by explorer endpoints.","type":"object","required":["data"],"properties":{"data":{"type":"array","items":{"$ref":"#/definitions/AuthorWithStats"}},"meta":{"default":{},"allOf":[{"$ref":"#/definitions/ResponseMeta"}]}}},"AuthorResponse":{"title":"ApiResponse_for_AuthorWithStats","description":"Standard response envelope used by explorer endpoints.","type":"object","required":["data"],"properties":{"data":{"$ref":"#/definitions/AuthorWithStats"},"meta":{"default":{},"allOf":[{"$ref":"#/definitions/ResponseMeta"}]}}},"AuthorThreadsResponse":{"title":"ApiResponse_for_Array_of_ThreadWithStarter","description":"Standard response envelope used by explorer endpoints.","type":"object","required":["data"],"properties":{"data":{"type":"array","items":{"$ref":"#/definitions/ThreadWithStarter"}},"meta":{"default":{},"allOf":[{"$ref":"#/definitions/ResponseMeta"}]}}},"AuthorWithStats":{"title":"AuthorWithStats","description":"Aggregated author statistics used in list and detail endpoints.","type":"object","required":["email","email_count","id","mailing_lists","name_variations","thread_count"],"properties":{"canonical_name":{"type":["string","null"]},"email":{"type":"string"},"email_count":{"type":"integer","format":"int64"},"first_email_date":{"type":["string","null"],"format":"date-time"},"first_seen":{"type":["string","null"],"format":"date-time"},"id":{"type":"integer","format":"int32"},"last_email_date":{"type":["string","null"],"format":"date-time"},"last_seen":{"type":["string","null"],"format":"date-time"},"mailing_lists":{"type":"array","items":{"type":"string"}},"name_variations":{"type":"array","items":{"type":"string"}},"thread_count":{"type":"integer","format":"int64"}}},"EmailHierarchy":{"title":"EmailHierarchy","description":"Email node enriched with depth information for thread rendering.","type":"object","required":["author_email","author_id","blob_oid","date","depth","id","mailing_list_id","message_id","subject"],"properties":{"author_email":{"type":"string"},"author_id":{"type":"integer","format":"int32"},"author_name":{"type":["string","null"]},"blob_oid":{"type":"string"},"body":{"type":["string","null"]},"created_at":{"type":["string","null"],"format":"date-time"},"date":{"type":"string","format":"date-time"},"depth":{"type":"integer","format":"int32"},"id":{"type":"integer","format":"int32"},"in_reply_to":{"type":["string","null"]},"mailing_list_id":{"type":"integer","format":"int32"},"message_id":{"type":"string"},"subject":{"type":"string"}}},"EmailList":{"title":"ApiResponse_for_Array_of_EmailWithAuthor","description":"Standard response envelope used by explorer endpoints.","type":"object","required":["data"],"properties":{"data":{"type":"array","items":{"$ref":"#/definitions/EmailWithAuthor"}},"meta":{"default":{},"allOf":[{"$ref":"#/definitions/ResponseMeta"}]}}},"EmailResponse":{"title":"ApiResponse_for_EmailWithAuthor","description":"Standard response envelope used by explorer endpoints.","type":"object","required":["data"],"properties":{"data":{"$ref":"#/definitions/EmailWithAuthor"},"meta":{"default":{},"allOf":[{"$ref":"#/definitions/ResponseMeta"}]}}},"EmailWithAuthor":{"title":"EmailWithAuthor","description":"Email row enriched with author metadata for API responses.","type":"object","required":["author_email","author_id","blob_oid","date","id","mailing_list_id","message_id","subject"],"properties":{"author_email":{"type":"string"},"author_id":{"type":"integer","format":"int32"},"author_name":{"type":["string","null"]},"blob_oid":{"type":"string"},"body":{"type":["string","null"]},"created_at":{"type":["string","null"],"format":"date-time"},"date":{"type":"string","format":"date-time"},"id":{"type":"integer","format":"int32"},"in_reply_to":{"type":["string","null"]},"mailing_list_id":{"type":"integer","format":"int32"},"message_id":{"type":"string"},"subject":{"type":"string"}}},"EnqueueRequest":{"title":"EnqueueRequest","type":"object","required":["payload"],"properties":{"max_attempts":{"default":null,"type":["integer","null"],"format":"int32"},"payload":true,"queue":{"default":"default","type":"string"},"run_at":{"default":null,"type":["string","null"],"format":"date-time"}}},"EnqueueResponse":{"title":"EnqueueResponse","type":"object","required":["id","queue","run_at","status"],"properties":{"id":{"type":"integer","format":"int64"},"queue":{"type":"string"},"run_at":{"type":"string","format":"date-time"},"status":{"type":"string"}}},"Job":{"title":"Job","description":"Background job record stored in Postgres.","type":"object","required":["attempts","created_at","id","max_attempts","payload","queue","run_at","status","updated_at"],"properties":{"attempts":{"type":"integer","format":"int32"},"created_at":{"type":"string","format":"date-time"},"id":{"type":"integer","format":"int64"},"last_error":{"type":["string","null"]},"locked_at":{"type":["string","null"],"format":"date-time"},"locked_by":{"type":["string","null"]},"max_attempts":{"type":"integer","format":"int32"},"payload":true,"queue":{"type":"string"},"run_at":{"type":"string","format":"date-time"},"status":{"type":"string"},"updated_at":{"type":"string","format":"date-time"}}},"JobList":{"title":"Array_of_Job","type":"array","items":{"$ref":"#/definitions/Job"}},"JobStats":{"title":"JobStats","description":"Summary statistics for the jobs table.","type":"object","required":["by_status","total"],"properties":{"by_status":{"type":"array","items":{"$ref":"#/definitions/StatusCount"}},"total":{"type":"integer","format":"int64"}}},"ListAggregateStats":{"title":"ListAggregateStats","description":"Aggregate mailing list statistics across the deployment.","type":"object","required":["totalAuthors","totalEmails","totalLists","totalThreads"],"properties":{"totalAuthors":{"type":"integer","format":"int64"},"totalEmails":{"type":"integer","format":"int64"},"totalLists":{"type":"integer","format":"int64"},"totalThreads":{"type":"integer","format":"int64"}}},"MailingList":{"title":"MailingList","description":"Mailing list record.","type":"object","required":["created_at","enabled","id","name","slug"],"properties":{"created_at":{"type":"string","format":"date-time"},"description":{"type":["string","null"]},"enabled":{"type":"boolean"},"id":{"type":"integer","format":"int32"},"last_synced_at":{"type":["string","null"],"format":"date-time"},"last_threaded_at":{"type":["string","null"],"format":"date-time"},"name":{"type":"string"},"slug":{"type":"string"}}},"MailingListList":{"title":"Array_of_MailingList","type":"array","items":{"$ref":"#/definitions/MailingList"}},"MailingListStats":{"title":"MailingListStats","description":"Summary statistics for a single mailing list.","type":"object","required":["authorCount","emailCount","threadCount"],"properties":{"authorCount":{"type":"integer","format":"int64"},"dateRangeEnd":{"type":["string","null"],"format":"date-time"},"dateRangeStart":{"type":["string","null"],"format":"date-time"},"emailCount":{"type":"integer","format":"int64"},"threadCount":{"type":"integer","format":"int64"}}},"MailingListStatsResponse":{"title":"ApiResponse_for_MailingListStats","description":"Standard response envelope used by explorer endpoints.","type":"object","required":["data"],"properties":{"data":{"$ref":"#/definitions/MailingListStats"},"meta":{"default":{},"allOf":[{"$ref":"#/definitions/ResponseMeta"}]}}},"PaginationMeta":{"title":"PaginationMeta","description":"Pagination metadata attached to list responses.","type":"object","required":["page","pageSize","totalItems","totalPages"],"properties":{"page":{"type":"integer","format":"int64"},"pageSize":{"type":"integer","format":"int64"},"totalItems":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int64"}}},"ResponseMeta":{"title":"ResponseMeta","description":"Metadata wrapper attached to explorer responses.","type":"object","properties":{"extra":{"type":["object","null"],"additionalProperties":true},"filters":{"type":["object","null"],"additionalProperties":true},"listId":{"type":["string","null"]},"pagination":{"anyOf":[{"$ref":"#/definitions/PaginationMeta"},{"type":"null"}]},"sort":{"type":"array","items":{"$ref":"#/definitions/SortDescriptor"}}}},"SortDescriptor":{"title":"SortDescriptor","description":"Sort descriptor returned in response metadata.","type":"object","required":["direction","field"],"properties":{"direction":{"$ref":"#/definitions/SortDirection"},"field":{"type":"string"}}},"SortDirection":{"title":"SortDirection","description":"Sort direction for list responses.","type":"string","enum":["asc","desc"]},"StatusCount":{"description":"Aggregate count per job status.","type":"object","required":["count","status"],"properties":{"count":{"type":"integer","format":"int64"},"status":{"type":"string"}}},"Thread":{"title":"Thread","description":"Thread metadata stored in the database.","type":"object","required":["id","last_date","mailing_list_id","root_message_id","start_date","subject"],"properties":{"id":{"type":"integer","format":"int32"},"last_date":{"type":"string","format":"date-time"},"mailing_list_id":{"type":"integer","format":"int32"},"message_count":{"type":["integer","null"],"format":"int32"},"root_message_id":{"type":"string"},"start_date":{"type":"string","format":"date-time"},"subject":{"type":"string"}}},"ThreadDetail":{"title":"ThreadDetail","description":"Thread details including the threaded list of emails.","type":"object","required":["emails","thread"],"properties":{"emails":{"type":"array","items":{"$ref":"#/definitions/EmailHierarchy"}},"thread":{"$ref":"#/definitions/Thread"}}},"ThreadDetailResponse":{"title":"ApiResponse_for_ThreadDetail","description":"Standard response envelope used by explorer endpoints.","type":"object","required":["data"],"properties":{"data":{"$ref":"#/definitions/ThreadDetail"},"meta":{"default":{},"allOf":[{"$ref":"#/definitions/ResponseMeta"}]}}},"ThreadList":{"title":"ApiResponse_for_Array_of_ThreadWithStarter","description":"Standard response envelope used by explorer endpoints.","type":"object","required":["data"],"properties":{"data":{"type":"array","items":{"$ref":"#/definitions/ThreadWithStarter"}},"meta":{"default":{},"allOf":[{"$ref":"#/definitions/ResponseMeta"}]}}},"ThreadWithStarter":{"title":"ThreadWithStarter","description":"Thread metadata augmented with starter author details.","type":"object","required":["id","last_date","mailing_list_id","root_message_id","start_date","starter_email","starter_id","subject"],"properties":{"id":{"type":"integer","format":"int32"},"last_date":{"type":"string","format":"date-time"},"mailing_list_id":{"type":"integer","format":"int32"},"message_count":{"type":["integer","null"],"format":"int32"},"root_message_id":{"type":"string"},"start_date":{"type":"string","format":"date-time"},"starter_email":{"type":"string"},"starter_id":{"type":"integer","format":"int32"},"starter_name":{"type":["string","null"]},"subject":{"type":"string"}}},"UpdateMailingListRequest":{"title":"UpdateMailingListRequest","type":"object","required":["enabled"],"properties":{"enabled":{"type":"boolean"}}},"WebhookEnqueueResponse":{"title":"WebhookEnqueueResponse","type":"object","required":["queued"],"properties":{"queued":{"type":"integer","format":"uint","minimum":0.0}}}}},"tags":[{"name":"Health","description":"Health and smoke test endpoints"},{"name":"Jobs","description":"Background job queue management"},{"name":"MailingLists","description":"Mailing list management endpoints"},{"name":"Webhooks","description":"Webhook triggers for sync pipelines"},{"name":"Lists","description":"Explorer list statistics endpoints"},{"name":"Emails","description":"Explorer email endpoints"},{"name":"Threads","description":"Explorer thread endpoints"},{"name":"Authors","description":"Explorer author endpoints"},{"name":"Docs","description":"OpenAPI documentation endpoints"}]}